chore(tests): change namespace
