+++ tests/Unit/Foundation/ValueType/EmailTypeTest.php
@@ -7,7 +7,7 @@ namespace App\Tests\Unit\Foundation\ValueType;
 use App\Foundation\Exception\InvalidArgumentException;
 use App\Foundation\ValueObject\Email;
 use App\Foundation\ValueType\EmailType;
-use App\Tests\Concern\WithFaker;
+use App\Tests\Concern\Util\WithFaker;
 use App\Tests\TestCase;
 
 class EmailTypeTest extends TestCase
