+++ tests/Unit/Siklid/Application/Auth/LoginSuccessHandlerTest.php
@@ -10,7 +10,7 @@ use App\Foundation\ValueObject\Username;
 use App\Siklid\Application\Auth\LoginSuccessHandler;
 use App\Siklid\Document\AccessToken;
 use App\Siklid\Document\User;
-use App\Tests\Concern\WithFaker;
+use App\Tests\Concern\Util\WithFaker;
 use App\Tests\TestCase;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\Request;
